fix(web): split web/stub iframe to avoid non-web compile; specify entrypoint and full-height; retry build
