+++ lib/iframe_stub.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/widgets.dart';
+
+class ArchiveIframe extends StatelessWidget {
+	const ArchiveIframe({super.key});
+
+	@override
+	Widget build(BuildContext context) {
+		return const SizedBox.shrink();
+	}
+}
+
