+++ lib/main.dart
@@ -1,9 +1,7 @@
-// ignore: avoid_web_libraries_in_flutter
-import 'dart:html' as html;
-// ignore: undefined_prefixed_name
-import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'iframe_stub.dart'
+	if (dart.library.html) 'iframe_web.dart' as iframe;
 
 void main() {
 	runApp(const TnblApp());
@@ -19,27 +17,10 @@ class TnblApp extends StatelessWidget {
 			debugShowCheckedModeBanner: false,
 			home: const Scaffold(
 				body: SafeArea(
-					child: _ArchiveIframe(),
+					child: iframe.ArchiveIframe(),
 				),
 			),
 		);
-	}
-}
-
-class _ArchiveIframe extends StatelessWidget {
-	const _ArchiveIframe();
-
-	@override
-	Widget build(BuildContext context) {
-		final element = html.IFrameElement()
-			..src = 'archive/index.html'
-			..style.border = '0'
-			..style.width = '100%'
-			..style.height = '100%';
-		html.document.body?.style.margin = '0';
-		// ignore: undefined_prefixed_name
-		ui.platformViewRegistry.registerViewFactory('iframe-view', (int viewId) => element);
-		return const HtmlElementView(viewType: 'iframe-view');
 	}
 }
